Move new agreements history tests to existing feature spec
Since there was duplication and they're part of the same workflow

Since there was a lot fo duplication and they're
